sort imports and minor refactors
