+++ app/containers/RoomItem/interfaces.ts
@@ -2,7 +2,7 @@ import React from 'react';
 import Animated from 'react-native-reanimated';
 
 import { TSupportedThemes } from '../../theme';
-import { TUserStatus, ILastMessage, SubscriptionType, IOmnichannelSource } from '../../definitions';
+import { TUserStatus, ILastMessage, SubscriptionType, IOmnichannelSource, ISubscription } from '../../definitions';
 
 export interface ILeftActionsProps {
 	transX: Animated.SharedValue<number>;
@@ -155,4 +155,9 @@ export interface IIconOrAvatar {
 	teamMain: boolean;
 	showLastMessage: boolean;
 	sourceType: IOmnichannelSource;
+}
+
+export interface IRoomItem extends ISubscription {
+	search?: boolean;
+	outside?: boolean;
 }
