+++ app/views/RoomsListView/components/Header.tsx
@@ -3,12 +3,12 @@ import { StyleSheet, Text, View, useWindowDimensions } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 
 import { showActionSheetRef } from '../../../containers/ActionSheet';
-import ServersList from './ServersList';
-import I18n from '../../../i18n';
-import sharedStyles from '../../Styles';
-import { useTheme } from '../../../theme';
 import SearchHeader from '../../../containers/SearchHeader';
+import I18n from '../../../i18n';
 import { useAppSelector } from '../../../lib/hooks';
+import { useTheme } from '../../../theme';
+import sharedStyles from '../../Styles';
+import ServersList from './ServersList';
 
 const styles = StyleSheet.create({
 	container: {
