--- app/views/RoomsListView/definitions.ts
@@ -1,6 +1,0 @@
-import { ISubscription } from '../../definitions';
-
-export interface IRoomItem extends ISubscription {
-	search?: boolean;
-	outside?: boolean;
-}
