+++ app/views/RoomsListView/hooks/useSubscriptions.ts
@@ -1,12 +1,12 @@
+import { Q } from '@nozbe/watermelondb';
 import { useEffect, useRef, useState } from 'react';
 import { shallowEqual } from 'react-redux';
-import { Q } from '@nozbe/watermelondb';
 import { Subscription } from 'rxjs';
 
+import { type TSubscriptionModel } from '../../../definitions';
 import { SortBy } from '../../../lib/constants';
 import database from '../../../lib/database';
 import { useAppSelector } from '../../../lib/hooks';
-import { TSubscriptionModel } from '../../../definitions';
 import { getUserSelector } from '../../../selectors/login';
 
 const CHATS_HEADER = 'Chats';
