+++ app/views/RoomsListView/index.tsx
@@ -8,6 +8,7 @@ import ActivityIndicator from '../../containers/ActivityIndicator';
 import BackgroundContainer from '../../containers/BackgroundContainer';
 import { ChangePasswordRequired } from '../../containers/ChangePasswordRequired';
 import RoomItem from '../../containers/RoomItem';
+import { type IRoomItem } from '../../containers/RoomItem/interfaces';
 import { SupportedVersionsExpired } from '../../containers/SupportedVersions';
 import i18n from '../../i18n';
 import { MAX_SIDEBAR_WIDTH } from '../../lib/constants';
@@ -19,13 +20,12 @@ import { getUserSelector } from '../../selectors/login';
 import { useTheme } from '../../theme';
 import Container from './components/Container';
 import ListHeader from './components/ListHeader';
-import { SectionHeader } from './components/SectionHeader';
-import { IRoomItem } from './definitions';
+import SectionHeader from './components/SectionHeader';
+import RoomsSearchProvider, { RoomsSearchContext } from './contexts/RoomsSearchProvider';
 import { useGetItemLayout } from './hooks/useGetItemLayout';
 import { useHeader } from './hooks/useHeader';
 import { useRefresh } from './hooks/useRefresh';
 import { useSubscriptions } from './hooks/useSubscriptions';
-import { RoomsSearchContext, RoomsSearchProvider } from './RoomsSearchProvider';
 import styles from './styles';
 
 const INITIAL_NUM_TO_RENDER = isTablet ? 20 : 12;
